pyproject.toml
[tool.poetry]
name = "ui-screenshot-to-prompt"
version = "0.3.0"
description = ""
authors = ["Stijn Smits"]
license = "MIT"
readme = "README.md"
package-mode = false  # 禁用打包模式

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.10.0.84"
gradio = "^5.4.0"
numpy = "^2.1.2"
pillow = "^11.0.0"
scikit-learn = "^1.5.2"
tqdm = "^4.66.5"
python-dotenv = "^1.0.1"
anthropic = "^0.37.1"
scikit-image = "^0.24.0"
openai = "^1.52.2"
pytesseract = "^0.3.13"
easyocr = "^1.7.2"
flask = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
